Scaffold slides
Using Remark for some nice slides, and starting to set up SystemJS to compile some Sass for a theme.

Note that because of how jspm currently works (designed for JS projects), the loader must explicitly be set when importing the Sass file, otherwise it would look for `theme.scss.js`.

Also, because we're serving slides out of a subdirectory, we'll need to update paths in the SystemJS config (which in this commit, was auto-generated by jspm). Changes to the SystemJS config will be in a follow-up commit, so that those changes are easier to see.